feat: create slash cmt for sale track
